chore: removed unused charts and deps
