+++ package.json
@@ -33,8 +33,6 @@
     "@mysten/sui": "^1.38.0",
     "@next/bundle-analyzer": "^15.5.2",
     "@next/third-parties": "^15.5.2",
-    "@nivo/core": "^0.99.0",
-    "@nivo/line": "^0.99.0",
     "@rive-app/react-canvas": "^4.23.0",
     "@safe-global/safe-apps-sdk": "^9.1.0",
     "@sentry/browser": "^9.42.1",
@@ -47,12 +45,10 @@
     "@strapi/blocks-react-renderer": "^1.0.2",
     "@tanstack/react-query": "^5.89.0",
     "@turtledev/api": "1.4.5",
-    "@types/d3-shape": "^3.1.7",
     "@wagmi/core": "^2.20.3",
     "@widgetbot/react-embed": "^1.9.0",
     "accept-language": "^3.0.20",
     "blo": "^1.2.0",
-    "d3-shape": "^3.2.0",
     "date-fns": "^4.1.0",
     "ethers": "^6.15.0",
     "framer-motion": "^11.18.2",
@@ -63,7 +59,6 @@
     "i18next-resources-for-ts": "^1.7.4",
     "i18next-resources-to-backend": "^1.2.1",
     "immer": "^10.1.3",
-    "lightweight-charts": "^5.0.8",
     "lodash": "^4.17.21",
     "next": "^15.5.2",
     "next-i18n-router": "^5.5.3",
