--- src/components/core/charts/LineChartV3/LineChartV3.tsx
@@ -1,160 +1,0 @@
-import { useTheme } from '@mui/material/styles';
-import {
-  AreaChart,
-  Area,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  ResponsiveContainer,
-} from 'recharts';
-import { CustomTooltip } from './CustomTooltip';
-import { toCompactValue } from 'src/utils/formatNumbers';
-import { LineChartV3Skeleton } from './LineChartV3Skeleton';
-import { calculateTooltipPosition } from './utils';
-import { useRef } from 'react';
-import { format } from 'date-fns';
-
-export interface ChartDataPoint<V> {
-  date: string;
-  value: V;
-}
-
-export interface LineChartV3Props<V, T extends ChartDataPoint<V>> {
-  data: T[];
-  theme: {
-    lineColor?: string;
-    areaTopColor?: string;
-    areaBottomColor?: string;
-    pointColor?: string;
-  };
-  dateFormat?: string;
-  dataSetId?: string;
-  isLoading?: boolean;
-  enableCrosshair?: boolean;
-  enableGridY?: boolean;
-  enableXAxis?: boolean;
-  enableYAxis?: boolean;
-  enableTooltip?: boolean;
-}
-
-export const LineChartV3 = <V, T extends ChartDataPoint<V>>({
-  data,
-  theme,
-  dateFormat,
-  dataSetId,
-  enableCrosshair = true,
-  enableGridY = true,
-  enableXAxis = true,
-  enableYAxis = true,
-  enableTooltip = true,
-  isLoading,
-}: LineChartV3Props<V, T>) => {
-  const muiTheme = useTheme();
-  const chartContainerRef = useRef<HTMLDivElement>(null);
-
-  if (isLoading) {
-    return <LineChartV3Skeleton />;
-  }
-
-  return (
-    <ResponsiveContainer ref={chartContainerRef} width="100%" height={300}>
-      <AreaChart data={data} accessibilityLayer={false}>
-        <defs>
-          <linearGradient id="areaGradient" x1="0" y1="0" x2="0" y2="1">
-            <stop offset="0%" stopColor={theme.areaTopColor} stopOpacity={1} />
-            <stop
-              offset="100%"
-              stopColor={theme.areaBottomColor}
-              stopOpacity={1}
-            />
-          </linearGradient>
-        </defs>
-        {enableGridY && (
-          <CartesianGrid
-            vertical={false}
-            horizontal={{
-              offset: -1,
-            }}
-            strokeDasharray="0"
-            stroke={`color-mix(in srgb, ${
-              (muiTheme.vars || muiTheme).palette.alpha900.main
-            } 10%, transparent)`}
-          />
-        )}
-        {enableXAxis && (
-          <XAxis
-            dataKey="date"
-            axisLine={false}
-            tickLine={false}
-            minTickGap={10}
-            tick={{
-              fill: (muiTheme.vars || muiTheme).palette.text.secondary,
-              fontSize: 10,
-              fontFamily: muiTheme.typography.bodyXXSmall.fontFamily,
-              fontWeight: muiTheme.typography.bodyXXSmall.fontWeight,
-            }}
-            tickFormatter={(value) => {
-              return format(value ?? '', dateFormat ?? 'MMM yyyy');
-            }}
-          />
-        )}
-        {enableYAxis && (
-          <YAxis
-            axisLine={false}
-            tickLine={false}
-            tickCount={5}
-            width={40}
-            tick={{
-              fill: (muiTheme.vars || muiTheme).palette.text.secondary,
-              fontSize: 10,
-              fontFamily: muiTheme.typography.bodyXXSmall.fontFamily,
-              fontWeight: muiTheme.typography.bodyXXSmall.fontWeight,
-            }}
-            tickFormatter={(value) => {
-              return !value || isNaN(Number(value))
-                ? value
-                : toCompactValue(value);
-            }}
-          />
-        )}
-        {enableTooltip && (
-          <Tooltip
-            content={(props) => {
-              const { x, y } = calculateTooltipPosition(
-                props.coordinate?.x ?? 0,
-                props.coordinate?.y ?? 0,
-                chartContainerRef.current?.clientWidth ?? 0,
-                chartContainerRef.current?.clientHeight ?? 0,
-              );
-              return (
-                <CustomTooltip
-                  {...props}
-                  x={x}
-                  y={y}
-                  dateFormat={dateFormat}
-                  dataSetId={dataSetId}
-                />
-              );
-            }}
-            cursor={false}
-          />
-        )}
-        <Area
-          type="natural"
-          dataKey="value"
-          stroke={theme.lineColor}
-          activeDot={
-            enableCrosshair
-              ? { r: 4, strokeWidth: 0, fill: theme.pointColor }
-              : false
-          }
-          fillOpacity={1}
-          fill="url(#areaGradient)"
-          isAnimationActive
-          baseValue={-1}
-        />
-      </AreaChart>
-    </ResponsiveContainer>
-  );
-};
